requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.0.8
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
redis==5.0.8
service-identity==24.1.0
setuptools==74.1.2
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
zope.interface==7.0.3
django-cors-headers
djangorestframework
djangorestframework-simplejwt